Word Search Problem
Iterate through the list, call the function identify if the char matches
that in the string. In the function,  if the char matches, then change
it state to be visited, and go on call the function on the four
sourrounding chars, if one of them match, go on, if all of them don't
match, return false and go back to last function stack and change the
char state to unvisited, which means it is not a part of the word and we
can check it later.